# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Life Circle Backend",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/analysis/hybrid": {
      "post": {
        "tags": [
          "Hybrid Isochrone"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_analysis_hybrid_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HybridRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/hybrid/by-request/{client_request_id}": {
      "get": {
        "tags": [
          "Hybrid Isochrone"
        ],
        "summary": "By Request",
        "operationId": "by_request_api_v1_analysis_hybrid_by_request__client_request_id__get",
        "parameters": [
          {
            "name": "client_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Client Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/analysis/hybrid/by-request/{client_request_id}/cancel": {
      "post": {
        "tags": [
          "Hybrid Isochrone"
        ],
        "summary": "Cancel By Request",
        "operationId": "cancel_by_request_api_v1_analysis_hybrid_by_request__client_request_id__cancel_post",
        "parameters": [
          {
            "name": "client_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Client Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/analysis/hybrid/{task_id}": {
      "get": {
        "tags": [
          "Hybrid Isochrone"
        ],
        "summary": "Status",
        "operationId": "status_api_v1_analysis_hybrid__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/analysis/hybrid/{task_id}/result": {
      "get": {
        "tags": [
          "Hybrid Isochrone"
        ],
        "summary": "Result",
        "operationId": "result_api_v1_analysis_hybrid__task_id__result_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridResultResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/analysis/hybrid/{task_id}/cancel": {
      "post": {
        "tags": [
          "Hybrid Isochrone"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_v1_analysis_hybrid__task_id__cancel_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FacilityAnalysis": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "failed"
            ],
            "title": "Status"
          },
          "queries": {
            "items": {
              "$ref": "#/components/schemas/QueryEvidence"
            },
            "type": "array",
            "title": "Queries"
          },
          "assessments": {
            "items": {
              "$ref": "#/components/schemas/AssessmentPoint"
            },
            "type": "array",
            "title": "Assessments"
          },
          "candidate_points": {
            "type": "integer",
            "title": "Candidate Points"
          },
          "assessed_points": {
            "type": "integer",
            "title": "Assessed Points"
          },
          "unassessed_points": {
            "type": "integer",
            "title": "Unassessed Points"
          },
          "network_requests": {
            "type": "integer",
            "title": "Network Requests"
          },
          "elapsed_seconds": {
            "type": "number",
            "title": "Elapsed Seconds"
          },
          "search_radius_m": {
            "type": "integer",
            "title": "Search Radius M"
          },
          "routes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RouteEvidence"
            },
            "type": "object",
            "title": "Routes"
          },
          "serviceBlindRegions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Geometry"
            },
            "propertyNames": {
              "enum": [
                "shopping",
                "medical",
                "education"
              ]
            },
            "type": "object",
            "title": "Serviceblindregions"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "queries",
          "assessments",
          "candidate_points",
          "assessed_points",
          "unassessed_points",
          "network_requests",
          "elapsed_seconds",
          "search_radius_m",
          "warnings"
        ],
        "title": "FacilityAnalysis"
      },
...
      "HybridError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "HybridError"
      },
...
      "HybridRequest": {
        "properties": {
          "origin": {
            "$ref": "#/components/schemas/Origin"
          },
          "coordinate_system": {
            "type": "string",
            "const": "bd09ll",
            "title": "Coordinate System"
          },
          "config": {
            "$ref": "#/components/schemas/HybridConfig"
          },
          "client_request_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Client Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "origin",
          "coordinate_system",
          "client_request_id"
        ],
        "title": "HybridRequest"
      },
      "HybridResultResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "responseType": {
            "type": "string",
            "const": "result",
            "title": "Responsetype",
            "default": "result"
          },
          "taskId": {
            "type": "string",
            "title": "Taskid"
          },
          "taskStatus": {
            "type": "string",
            "const": "completed",
            "title": "Taskstatus"
          },
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "failed",
              "empty"
            ],
            "title": "Status"
          },
          "businessStatus": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "failed",
              "empty"
            ],
            "title": "Businessstatus"
          },
          "dataSource": {
            "type": "string",
            "enum": [
              "synthetic",
              "baidu_walking"
            ],
            "title": "Datasource"
          },
          "center": {
            "$ref": "#/components/schemas/Origin"
          },
          "generatedAt": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Generatedat"
          },
          "facilitiesStatus": {
            "type": "string",
            "enum": [
              "not_integrated",
              "complete",
              "partial",
              "failed"
            ],
            "title": "Facilitiesstatus"
          },
          "facilityAnalysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FacilityAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "coordinateSystem": {
            "type": "string",
            "const": "bd09ll",
            "title": "Coordinatesystem",
            "default": "bd09ll"
          },
          "coordinateOrder": {
            "type": "string",
            "const": "longitude,latitude",
            "title": "Coordinateorder",
            "default": "longitude,latitude"
          },
          "units": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Units"
          },
          "rules": {
            "$ref": "#/components/schemas/Rules"
          },
          "data": {
            "$ref": "#/components/schemas/Data"
          },
          "algorithm": {
            "$ref": "#/components/schemas/HybridIsochrone"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "type": "array",
            "title": "Warnings"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "type": "array",
            "title": "Errors"
          },
          "isochrone": {
            "$ref": "#/components/schemas/HybridIsochrone"
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "result_hash": {
            "type": "string",
            "title": "Result Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "taskId",
          "taskStatus",
          "status",
          "businessStatus",
          "dataSource",
          "center",
          "generatedAt",
          "facilitiesStatus",
          "algorithm",
          "isochrone",
          "config_hash",
          "result_hash"
        ],
        "title": "HybridResultResponse"
      },
...
      "Origin": {
        "properties": {
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng"
          },
          "lat": {
            "type": "number",
            "exclusiveMaximum": 85.0,
            "exclusiveMinimum": -85.0,
            "title": "Lat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lng",
          "lat"
        ],
        "title": "Origin"
      },
...
      "Rules": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version",
            "default": "n04-v1"
          },
          "time_threshold_s": {
            "type": "integer",
            "const": 900,
            "title": "Time Threshold S",
            "default": 900
          },
          "time_inclusive": {
            "type": "boolean",
            "const": true,
            "title": "Time Inclusive",
            "default": true
          },
          "time_tolerance_s": {
            "type": "number",
            "title": "Time Tolerance S",
            "default": 0
          },
          "time_confirmation": {
            "type": "string",
            "enum": [
              "implementation_only",
              "mock_only",
              "baidu_sampled"
            ],
            "title": "Time Confirmation",
            "default": "implementation_only"
          },
          "distance": {
            "$ref": "#/components/schemas/DistanceRule"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Rules"
      },
...
      "TaskStatusResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "responseType": {
            "type": "string",
            "const": "task",
            "title": "Responsetype",
            "default": "task"
          },
          "taskId": {
            "type": "string",
            "title": "Taskid"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "cancelling",
              "completed",
              "cancelled",
              "failed"
            ],
            "title": "Status"
          },
          "businessStatus": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "complete",
                  "partial",
                  "failed",
                  "empty"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Businessstatus"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "requests": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Requests"
          },
          "networkRequests": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Networkrequests"
          },
          "budget": {
            "type": "integer",
            "maximum": 800.0,
            "minimum": 1.0,
            "title": "Budget"
          },
          "elapsedSeconds": {
            "type": "number",
            "minimum": 0.0,
            "title": "Elapsedseconds"
          },
          "dataSource": {
            "type": "string",
            "enum": [
              "synthetic",
              "baidu_walking"
            ],
            "title": "Datasource"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "taskId",
          "status",
          "stage",
          "requests",
          "networkRequests",
          "budget",
          "elapsedSeconds",
          "dataSource"
        ],
        "title": "TaskStatusResponse",
        "description": "统一异步任务响应；business_status 是业务结果，status 是传输任务状态。"
      }
    }
  }
}